requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<1.0
django-filter>=24.1,<25.0
drf-spectacular>=0.27,<1.0
django-axes>=6.4,<7.0
celery>=5.3,<6.0
redis>=5.0,<6.0
django-redis>=5.4,<6.0
httpx>=0.27,<1.0
psycopg2-binary>=2.9,<3.0
whitenoise>=6.6,<7.0
daphne>=4.1,<5.0
